+++ modules/generator/registry/overrides.go
@@ -9,6 +9,7 @@ import (
 
 type Overrides interface {
 	MetricsGeneratorMaxActiveSeries(userID string) uint32
+	MetricsGeneratorMaxActiveEntities(userID string) uint32
 	MetricsGeneratorCollectionInterval(userID string) time.Duration
 	MetricsGeneratorDisableCollection(userID string) bool
 	MetricsGeneratorGenerateNativeHistograms(userID string) histograms.HistogramMethod