+++ modules/overrides/interface.go
@@ -48,6 +48,7 @@ type Interface interface {
 	MetricsGeneratorRingSize(userID string) int
 	MetricsGeneratorProcessors(userID string) map[string]struct{}
 	MetricsGeneratorMaxActiveSeries(userID string) uint32
+	MetricsGeneratorMaxActiveEntities(userID string) uint32
 	MetricsGeneratorCollectionInterval(userID string) time.Duration
 	MetricsGeneratorDisableCollection(userID string) bool
 	MetricsGeneratorGenerateNativeHistograms(userID string) histograms.HistogramMethod